Resolve all eslint warnings
