+++ airflow-core/src/airflow/ui/src/components/ui/Menu/index.ts
@@ -17,6 +17,4 @@
  * under the License.
  */
 
-export * from "./TimezoneProvider";
-export * from "./useTimezone";
-export * from "./Context";
+export * from "./Menu";
