+++ airflow-core/src/airflow/ui/src/components/ui/Toaster/index.ts
@@ -17,6 +17,5 @@
  * under the License.
  */
 
-export * from "./TimezoneProvider";
-export * from "./useTimezone";
-export * from "./Context";
+export * from "./Toaster";
+export * from "./createToaster";
