+++ airflow-core/src/airflow/ui/src/context/openGroups/index.ts
@@ -19,3 +19,4 @@
 
 export * from "./OpenGroupsProvider";
 export * from "./useOpenGroups";
+export * from "./Context";
